refactor(imagetool): split viewer internals into focused modules
