+++ src/erlab/interactive/imagetool/__init__.py
@@ -9,6 +9,7 @@ Modules
    :toctree:
 
    core
+   viewer
    slicer
    fastslicing
    fastbinning
